requirements.txt
tweepy==11.14.0
python-dotenv==1.0.0
